+++ navigator_grade/__init__.py
@@ -2,6 +2,8 @@
 
 from .calculator import (
     CompetencyRecord,
+    FutureOutlook,
+    FuturePlan,
     GradeAnalysis,
     UpgradeOption,
     acquisition_percentage,
@@ -9,6 +11,7 @@ from .calculator import (
     average_competency_level,
     grade_requirements_met,
     highest_grade_met,
+    future_outlook,
 )
 from .rules import GRADE_REQUIREMENTS, GradeRequirement
 
@@ -16,6 +19,8 @@ __all__ = [
     "GRADE_REQUIREMENTS",
     "GradeRequirement",
     "CompetencyRecord",
+    "FutureOutlook",
+    "FuturePlan",
     "GradeAnalysis",
     "UpgradeOption",
     "acquisition_percentage",
@@ -23,4 +28,5 @@ __all__ = [
     "average_competency_level",
     "grade_requirements_met",
     "highest_grade_met",
+    "future_outlook",
 ]